DOCSP-4552 Add code example pasting doclet
Adds a doclet (javadoc 'plugin') that generates a manifest of which
external code examples are available for the entities.

Adds a python script using BeautifulSoup to edit the resulting HTML
output of javadoc to inject the code examples.

Adds a couple code examples for starters (borrowed from the main docs
site). To come: testing code examples and just more code examples,
plus styling of the overall javadoc output.

Finally, it adds a generate_docs.sh script, similar to JS SDK.

A note on this approach:

From what I can tell, it is not possible to leverage the existing
javadoc HTML doclet implementation while modifying it as HtmlDoclet
is private to the JDK.

It's also not possible to modify the entities passed to my doclet,
e.g. to inject new code example tags. The entities are completely
immutable when passed to my doclet.

Failing all that, I thought maybe I could override the overall
DocletEnvironment context to return my own brand new tree of entities
with tags injected -- i.e. use the decorator pattern to pass a custom
DocletEnvironment implementation that wraps their implementation
to the StandardDoclet. Unfortunately, the JDK downcasts to their
concrete DocEnvImpl type somewhere deep in the code, which makes
that approach also impossible.

So, here we are at programmatically editing the HTML after the javadoc
step has completed.